requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
scikit-image>=0.21.0
tqdm>=4.65.0
pyyaml>=6.0
